refactor: move song minting logic into separate component
Handling of song minting in the hottestSongs page is of no relevance
because the page itself has no use of the states (showModal, isVisible)
the handle function manages. Furthermore, all the external calls the
functions makes is being imported from a separate module, hence making
th logic easy to remove from the page.

Seeing as MintSongButton and MintSongModal are the only components sharing
and managing the aforementioned states and functions, they got moved into
a shared parent component MintSong which will handle the states and functions
for it's children (MintSongButton and MintSongModal) which solves unnecessary
re-rendering of other components inside the hottestSongs page that don't need
the state